feat: Extracting labelsDrawer as a public component
