+++ web/src/pages/Route/typing.d.ts
@@ -106,16 +106,6 @@ declare namespace RouteModule {
     key: string;
   };
 
-  type ResponseLabelList = Record<string, string>[];
-
-  type LabelList = Record<string, string[]>;
-
-  type LabelTableProps = {
-    labelKey: string;
-    labelValue: string;
-    key: string;
-  };
-
   type Step1PassProps = {
     form: FormInstance;
     advancedMatchingRules: MatchingRule[];
